Rename resource type
