+++ src/extension.ts
@@ -5,8 +5,8 @@ import { write } from './components/clipboard';
 import { multiStepInput } from './config_file/action.addprofile';
 import { OscExplorer } from "./explorer";
 import { ExplorerResourceNode } from './flat/node';
-import { ResourceNode } from './flat/node.resources';
-import { VmResourceNode } from './flat/node.resources.vms';
+import { ResourceNode } from './flat/resources/node.resources';
+import { VmResourceNode } from './flat/resources/node.resources.vms';
 import { OscVirtualContentProvider } from './virtual_filesystem/oscvirtualfs';
 import { LogsProvider } from './virtual_filesystem/logs';
 import { ProfileNode } from './flat/node.profile';
